requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
array-record==0.4.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
blobfile==2.1.1
cachetools==5.3.1
certifi==2024.8.30
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
datasets==3.1.0
debugpy==1.8.12
decorator==5.2.0
dill==0.3.8
dm-tree==0.1.8
etils==1.3.0
executing==2.2.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.0
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
gviz-api==1.10.0
h5py==3.9.0
huggingface-hub==0.29.1
idna==3.4
importlib-resources==6.1.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.2
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lmdb==1.6.2
lxml==4.9.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.7
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
ml-collections==0.1.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.0
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.20.0
Pygments==2.19.1
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.2.1
regex==2024.5.10
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-profile==2.15.0
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-gan==2.1.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.21.0
termcolor==2.3.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.1.2+cu118
torchaudio==0.7.0a0+ac17b64
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
wcwidth==0.2.13
Werkzeug==2.3.7
wrapt==1.15.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
